scoperobit/requirements.txt
markovify
django-bootstrap4
